sha2算法更直观

+++ Encryption-Algorithm/encrypt/sha.h
@@ -6,13 +6,13 @@
 void sha1Encode(const unsigned char* message, unsigned long long messageLen, unsigned char* out);
 
 #define SHA224_OUTLEN 28
-void sha224Encode(const unsigned char* message, int messageLen, unsigned char* out);
+void sha224Encode(const unsigned char* message, unsigned long long messageLen, unsigned char* out);
 
 #define SHA256_OUTLEN 32
-void sha256Encode(const unsigned char* message, int messageLen, unsigned char* out);
+void sha256Encode(const unsigned char* message, unsigned long long messageLen, unsigned char* out);
 
 #define SHA384_OUTLEN 48
-void sha384Encode(const unsigned char* message, int messageLen, unsigned char* out);
+void sha384Encode(const unsigned char* message, unsigned long long messageLen, unsigned char* out);
 
 #define SHA512_OUTLEN 64
-void sha512Encode(const unsigned char* message, int messageLen, unsigned char* out);
+void sha512Encode(const unsigned char* message, unsigned long long messageLen, unsigned char* out);
